Split I/O into data sources and results handling
